Capstone2/requirements.txt
torch
torchvision
fastapi
uvicorn
pillow
numpy
evidently
pandas